requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
dj-static==0.0.6
Django==3.2.4
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-crispy-forms==1.12.0
django-datetime-widget2==0.9.5
django-extensions==3.1.3
django-filter==2.4.0
django-widget-tweaks==1.4.8
future==0.18.2
html5lib==1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
kombu==5.1.0
moviepy==1.0.3
numpy==1.21.0
oauthlib==3.1.1
Pillow==8.2.0
proglog==0.1.9
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
reportlab==3.6.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
static-ranges==0.2.0
static3==0.7.0
tqdm==4.61.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
